audio: Avoid mixer path dependency conflicts
On some devices, certain input and output mixer controls share the same
endpoint, which means that if an input stream is disabled (and the mixer
is updated), that shared endpoint may be disabled as well.

This would result in no audio, because no one is aware that the output
stream in fact does still need the shared endpoint to be enabled.

We prevent this by enabling the output mixers after disabling input mixers,
if any.

Change-Id: I4d80fcd0648a11099f27177da3cd29095b6a8c34
(cherry picked from commit 507e8ce5bffb36fae8b1cd4e5498fa19a6273ee6)